Cleared all known Bugs

+++ dataprovider/delete-analysedArchives.js
@@ -5,6 +5,7 @@ var helper = require('./twapper_modules/helper.js');
 
 var conn;
 var db;
+var laids =["29863de6315290d576d34e93d122c944-77","29863de6315290d576d34e93d122c944-75"];
 helper.getDBConnectionFromConfig(conf.couchdb, function(error, database, connection){
   if(error == null){
     conn = connection;
@@ -14,8 +15,17 @@ helper.getDBConnectionFromConfig(conf.couchdb, function(error, database, connect
       
       _.each(res, function(entry){
         if(entry.id !== "config" && entry.id.indexOf("_design") == -1){
-          db.remove(entry.id, entry.value.rev)
-          console.log("removed: ",entry);
+          if(laids.length > 0){
+            _.each(laids, function(laid){
+              if(entry.id.indexOf(laid) != -1){
+                db.remove(entry.id, entry.value.rev)
+                console.log("removed: ",entry);
+              }
+            });
+          }else{
+            db.remove(entry.id, entry.value.rev)
+            console.log("removed: ",entry);
+          }
         }
         
       });
